Working!
Will need to add readme, but this is a good stopping point. Checkout virtual-boy.py on an example of how this works. First, we create the prices file, then we pull the game IDs into a txt file, then we do a call for each one and parse that info and spit it out.